mongo-dump-backend/requirements.txt
fastapi
uvicorn[standard]
pymongo
motor
pydantic[email]
python-dotenv
bcrypt
python-jose[cryptography]
redis
